tickets/3496: int32 paymentauthorisationId to int64

+++ AutonomousProviderService/AutonomousProviderService.svc.cs
@@ -80,7 +80,7 @@ namespace AutonomousProviderService
                     res.TokenType = dr["TOKENTYPE"].ToString();
 
                     if (string.IsNullOrEmpty(res.Remark)) {
-                        int paymentAuthorisationId = Functions.GenerateUniqueId();
+                        ulong paymentAuthorisationId = Functions.GenerateUniqueId();
                         res.PaymentAuthorisationId = paymentAuthorisationId.ToString();
                         CreateRegistration(req, paymentAuthorisationId);
                     }
@@ -168,7 +168,7 @@ namespace AutonomousProviderService
         }
         #endregion
 
-        private void CreateRegistration(PaymentStartRequest req, int AuthorisationId)
+        private void CreateRegistration(PaymentStartRequest req, ulong AuthorisationId)
         {
             if (ConfigurationManager.AppSettings["AVG"] == null || !bool.Parse(ConfigurationManager.AppSettings["AVG"])) return;
