+++ VerwijsIndexInterface/Functions.cs
@@ -107,10 +107,10 @@ namespace Denion.WebService.VerwijsIndex
             return sb.ToString();
         }
 
-        public static int GenerateUniqueId() {
+        public static ulong GenerateUniqueId() {
             var now = DateTime.Now;
             var zeroDate = DateTime.MinValue.AddHours(now.Hour).AddMinutes(now.Minute).AddSeconds(now.Second).AddMilliseconds(now.Millisecond);
-            return (int)(zeroDate.Ticks / 10000) % int.MaxValue-1;
+            return (ulong)(zeroDate.Ticks / 10000) % int.MaxValue-1;
         }
     }
